Add Smooch JavaScript
The initialization code (provided by Smooch) is now added during asset
processing. This includes adding a global JavaScript variable containing
the Smooch app ID from the plugin configuration so that it can be passed
through to the Smooch initialization code.

With this change in place, the plugin is now functioning and displays
the live chat window.